Add tests for record field code actions
